// src/api/triton/types.gen.ts
export type TaskStatus = 'queued' | 'downloading' | 'processing' | 'completed' | 'failed';

/**
 * TaskType
 */
export type TaskType = 'youtube' | 'podcast' | 'video' | 'audio' | 'pdf' | 'image';